refactor: 컨트롤러를 분리하고 조건 분기문을 interface 를 사용하여 제거
