+++ src/main/java/Application.java
@@ -1,7 +1,7 @@
-import controller.PosController;
+import controller.PosMachine;
 
 public class Application {
     public static void main(String[] args) {
-        PosController.run();
+        PosMachine.run();
     }
 }
