+++ src/main/java/controller/TerminateController.java
@@ -1,0 +1,11 @@
+package controller;
+
+import domain.Menus;
+import domain.Tables;
+
+public class TerminateController implements PosController {
+    @Override
+    public void execute(Tables tables, Menus menus) {
+        return;
+    }
+}
